+++ pycognaize/common/enums.py
@@ -91,6 +91,7 @@ class IqFieldKeyEnum(enum.Enum):
     tags = 'tags'
     template_id = 'templateId'
     value = 'value'
+    calculated_value = 'calculatedValue'
     ocr_value = 'ocrValue'
     repeat_parent = 'repeatParent'
     section = 'section'
@@ -224,3 +225,19 @@ class XBRLTableTagEnum(enum.Enum):
     cells = 'cells'
     title = 'title'
     _id = '_id'
+
+
+class PageLayoutEnum(enum.Enum):
+    """Enum of page layout blocks, including tables"""
+    TEXT = 'page_layout__text'
+    # 'page_layout__picture'
+    CAPTION = 'page_layout__caption'
+    SECTION_HEADER = 'page_layout__section_header'
+    FOOTNOTE = 'page_layout__footnote'
+    FORMULA = 'page_layout__formula'
+    # 'page_layout__table'
+    LIST_ITEM = 'page_layout__list_item'
+    PAGE_HEADER = 'page_layout__page_header'
+    PAGE_FOOTER = 'page_layout__page_footer'
+    TITLE = 'page_layout__title'
+    TABLE = 'tables__table'