+++ pycognaize/document/tag/html_tag.py
@@ -1,9 +1,9 @@
 import abc
-from typing import Optional, List, Union
-
 import bson
 import pandas as pd
+import logging
 
+from typing import Optional, List, Union
 from pycognaize.common.enums import (IqRecipeEnum, XBRLCellEnum,
                                      XBRLTableTagEnum, XBRLTagEnum, ID)
 from pycognaize.document.html_info import HTML
@@ -207,6 +207,30 @@ class HTMLTableTag(HTMLTagABC):
                                                tag_id=self.tag_id,
                                                row_index=row_index,
                                                col_index=col_index)
+        df = self.replace_nans_with_empty_html_tags(df)
+        return df
+
+    def replace_nans_with_empty_html_tags(self,
+                                          df: pd.DataFrame) -> pd.DataFrame:
+        """
+            Replaces NaN values in a DataFrame with empty HTML tags.
+        """
+        for col in df.columns:
+            for idx in df.index:
+                if pd.isna(df.loc[idx, col]):
+                    logging.warning(
+                        f'Build df issue: Replacing empty cell at {idx, col} '
+                        f'with empty HTMLTag in HTMLTableTag with html id '
+                        f'{self.html_id}')
+                    df.loc[idx, col] = HTMLTag(is_table=False,
+                                               html_id='',
+                                               xpath='',
+                                               raw_value='',
+                                               raw_ocr_value='',
+                                               field_id='',
+                                               tag_id='',
+                                               row_index=idx,
+                                               col_index=col)
         return df